Remove orphan arbitrary instances
